delete clubs.tsx, change clubsnew to clubs
